extract API call from created fetch directory for all api calls
